refactor(v0.80.0): project_detect.py core/cli → core/utils 이전 (의존성 정리 A3, 5/5)
새 사이클 A3. v0.52에서 deferred됐던 33개 ignore_imports 중 추가 1개 정리.

이전: core/cli/project_detect.py (377L) → core/utils/project_detect.py
- 프로젝트 타입 + 10개 AI 하니스 디렉토리 자동 탐지 — 순수 path inspection
- CLI 의존성 0. v0.40 init 명령 도입 시기의 historical placement

호출자 업데이트 (4 파일):
- core/memory/context.py:292 (lazy import — layer 위반이었음)
- core/cli/welcome.py:33
- core/cli/typer_init.py:50
- tests/test_project_detect.py:7

pyproject.toml ignore_imports 1 entry 제거:
- "core.memory.context -> core.cli.project_detect" (Server contract)

29 → 28 ignore_imports 남음.

Quality Gates:
- ruff/ruff-format/mypy 0 errors (329 source files)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- import-linter 4/4 contracts 깨짐 0
- E2E Cowboy Bebop A (68.4) 변동 없음
- ruff I001 1건 auto-fix

5 PR 진척:
- A1 redaction → utils ✅ v0.78.0
- A2 bash_tool → agent ✅ v0.79.0
- A3 (this) project_detect → utils ✅ v0.80.0
- A4 session_checkpoint + transcript → runtime_state/
- A5 startup → lifecycle
