+++ core/utils/project_detect.py
