+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geode"
-version = "0.79.0"
+version = "0.80.0"
 description = "GEODE — 범용 자율 실행 에이전트"
 license = "Apache-2.0"
 readme = "README.md"
@@ -121,7 +121,6 @@ ignore_imports = [
     "core.server.supervised.services -> core.cli.tool_handlers",
     "core.server.supervised.services -> core.cli.bootstrap",
     "core.lifecycle.bootstrap -> core.cli.startup",
-    "core.memory.context -> core.cli.project_detect",
     "core.agent.worker -> core.cli.tool_handlers",
     # Tier 3 #7 — loop.py was split into core/agent/loop/ package
     "core.agent.loop.loop -> core.cli.session_checkpoint",
